Remove getWindowLocationSearch global and duplicate implementation
